+++ src/models/SearchTab.ts
@@ -1,0 +1,18 @@
+import { ArchivingListContent } from './Archiving'
+
+export interface SearchTabData {
+  searchData: {
+    archivings: {
+      content: ArchivingListContent[]
+      page: number
+      size: number
+      hasNext: boolean
+    }
+    community: {
+      content: ArchivingListContent[]
+      page: number
+      size: number
+      hasNext: boolean
+    }
+  }
+}
