+++ src/models/enums/SearchType.ts
@@ -1,0 +1,5 @@
+export enum SearchType {
+  All = 'ALL',
+  My = 'MY',
+  Community = 'COMMUNITY',
+}
